requirements.txt
numpy==1.26.3
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
python_pcl==0.3
